requirements.txt
python-telegram-bot==21.6
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.3.0
langchain-chroma>=0.1.0
langchain-gigachat>=0.3.0
langchain-text-splitters>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pymupdf==1.24.9
selenium==4.23.1
tiktoken==0.7.0
python-dotenv==1.0.1
requests>=2.31.0
urllib3>=2.0.0

